add: authentication

+++ backend/src/controller/StudentController.js
@@ -1,8 +1,9 @@
 const StudentService = require("../service/StudentService");
 
 module.exports = {
-  async findAll(_, response){
+  async findAll(req, response){
     try{
+      if(!req.user.isAdmin) throw new Error('Não autorizado')
         const students = await StudentService.findAll();
         return response.json(students);
     }catch(e){
@@ -14,6 +15,7 @@ module.exports = {
 
   async createStudent(request, response) {
     try{
+      if(!request.user.isAdmin) throw new Error('Não autorizado')
       const { ra, cpf, email, name } = request.body;
       const result = await StudentService.createStudent( ra, cpf, email, name )
       return response.json(result);
@@ -26,6 +28,7 @@ module.exports = {
   
   async destroyStudent(request, response){
     try{
+      if(!request.user.isAdmin) throw new Error('Não autorizado')
       const { ra } = request.params;
       const result = await StudentService.deleteStudent(ra);
       return response.json(result);
@@ -38,6 +41,7 @@ module.exports = {
 
   async getStudentByRA(request, response){
     try{
+      if(!request.user.isAdmin) throw new Error('Não autorizado')
       const { ra } = request.params;
       const result = await StudentService.findStudentByRA(ra)
       return response.json(result);
@@ -50,6 +54,7 @@ module.exports = {
 
   async filterStudentByName(request, response){
     try{
+      if(!request.user.isAdmin) throw new Error('Não autorizado')
       const { name } = request.params;
       const result = await StudentService.filterStudentByName(name)
       return response.json(result);
@@ -62,6 +67,7 @@ module.exports = {
 
   async updateStudent(request, response){
     try{
+      if(!request.user.isAdmin) throw new Error('Não autorizado')
       const { ra } = request.params;
       const { email, name } = request.body;
       const result = await StudentService.updateStudent({name, email}, ra)
